split test files and add global integration test setup
